+++ Backend/src/main/java/app/controller/ClientController.java
@@ -1,7 +1,12 @@
 package app.controller;
 
+import app.entity.UserSession;
+import app.entity.user.User;
+import app.entity.user.UserRole;
 import app.infraestructure.Response;
 import app.model.ClientModel;
+import app.repository.UserRepository;
+import app.repository.UserSessionRepository;
 import app.service.ClientService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -17,10 +22,24 @@ public class ClientController {
     @Autowired
     private ClientService clientService;
 
+    @Autowired
+    UserSessionRepository sessionRepository;
+
+    @Autowired
+    UserRepository _userRepository;
+
     @PostMapping(value = "/client")
     @ResponseBody
-    public ResponseEntity<Response> createClient(@RequestBody ClientModel newClient){
-        Response response = clientService.createClient(newClient);
+    public ResponseEntity<Response> createClient(@RequestBody ClientModel newClient ,@RequestHeader(value = "Authorization") String sessionToken){
+
+        Response response = new Response();
+
+        UserSession session = sessionRepository.getByToken(sessionToken);
+        User user = session.getUserId();
+
+        if (user != null && user.getRole().equals(UserRole.ADMINISTRATOR.getCode())) {
+            response = clientService.createClient(newClient);
+        } else response.addError("El usuario no es administrador");
 
         if(response.hasErrors()) return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
 
@@ -29,8 +48,17 @@ public class ClientController {
 
     @DeleteMapping(value = "/client/{clientId}")
     @ResponseBody
-    public ResponseEntity<Response> deleteClient(@PathVariable("clientId") long clientId){
-        Response response = clientService.deleteClient(clientId);
+    public ResponseEntity<Response> deleteClient(@PathVariable("clientId") long clientId,@RequestHeader(value = "Authorization") String sessionToken){
+
+        Response response = new Response();
+
+        UserSession session = sessionRepository.getByToken(sessionToken);
+        User user = session.getUserId();
+
+        if (user != null && user.getRole().equals(UserRole.ADMINISTRATOR.getCode())) {
+            response = clientService.deleteClient(clientId);
+        } else response.addError("El usuario no es administrador");
+
 
         if(response.hasErrors()) return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
 
@@ -39,8 +67,17 @@ public class ClientController {
 
     @PutMapping(value = "/client")
     @ResponseBody
-    public ResponseEntity<Response> updateClient(@RequestBody ClientModel newClient){
-        Response response = clientService.updateClient(newClient);
+    public ResponseEntity<Response> updateClient(@RequestBody ClientModel newClient,@RequestHeader(value = "Authorization") String sessionToken){
+
+        Response response = new Response();
+
+        UserSession session = sessionRepository.getByToken(sessionToken);
+        User user = session.getUserId();
+
+        if (user != null && user.getRole().equals(UserRole.ADMINISTRATOR.getCode())) {
+            response = clientService.updateClient(newClient);
+        } else response.addError("El usuario no es administrador");
+
 
         if(response.hasErrors()) return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
 
@@ -49,8 +86,16 @@ public class ClientController {
 
     @GetMapping(value = "/client")
     @ResponseBody
-    public ResponseEntity<Response> getClients(){
-        Response response = clientService.getAllClients();
+    public ResponseEntity<Response> getClients(@RequestHeader(value = "Authorization") String sessionToken){
+
+        Response response = new Response();
+
+        UserSession session = sessionRepository.getByToken(sessionToken);
+        User user = session.getUserId();
+
+        if (user != null && user.getRole().equals(UserRole.ADMINISTRATOR.getCode())) {
+            response = clientService.getAllClients();
+        } else response.addError("El usuario no es administrador");
 
         if(response.hasErrors()) return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
 
@@ -59,8 +104,17 @@ public class ClientController {
 
     @GetMapping(value = "/client/{clientId}")
     @ResponseBody
-    public ResponseEntity<Response> getById(@PathVariable("clientId") long idClient){
-        Response response = clientService.getById(idClient);
+    public ResponseEntity<Response> getById(@PathVariable("clientId") long idClient,@RequestHeader(value = "Authorization") String sessionToken){
+        Response response = new Response();
+
+        UserSession session = sessionRepository.getByToken(sessionToken);
+        User user = session.getUserId();
+
+        if (user != null && user.getRole().equals(UserRole.ADMINISTRATOR.getCode())) {
+
+            response = clientService.getClientById(idClient);
+
+        } else response.addError("El usuario no es administrador");
 
         if(response.hasErrors()) return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
 
@@ -69,8 +123,19 @@ public class ClientController {
 
     @GetMapping(value = "/client-by-cuit")
     @ResponseBody
-    public ResponseEntity<Response> getByCuit(@RequestParam("cuit") long cuit){
-        Response response = clientService.getByCuit(cuit);
+    public ResponseEntity<Response> getByCuit(@RequestParam("cuit") long cuit,@RequestHeader(value = "Authorization") String sessionToken){
+
+
+        Response response = new Response();
+
+        UserSession session = sessionRepository.getByToken(sessionToken);
+        User user = session.getUserId();
+
+        if (user != null && user.getRole().equals(UserRole.ADMINISTRATOR.getCode())) {
+
+            response = clientService.getByCuit(cuit);
+
+        } else response.addError("El usuario no es administrador");
 
         if(response.hasErrors()) return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
